Add organized folder structure skeleton
- Add folder structure skeleton
- Move some files according to new folder structure
