+++ src/App.tsx
@@ -2,7 +2,7 @@ import React, { useReducer, useRef, useState, useEffect, createContext } from "r
 import { DragDropContext, Droppable, DropResult } from "react-beautiful-dnd";
 
 import InputField from "./components/InputField/InputField";
-import { Actions, Todo, TodoListsType, SortingStatusType, tabSearchInputsType } from "./model";
+import { Actions, Todo, TodoListsType, SortingStatusType, tabSearchInputsType } from "./models/model";
 import TodoList from "./components/TodosList/TodoList";
 import Footer from "./components/Footer/Footer";
 import "./App.scss";
