+++ src/data/Projects.ts
@@ -1,0 +1,1 @@
+export {}
