+++ src/pages/Home/Home.tsx
@@ -1,0 +1,1 @@
+export {}
